// src/app/core/enums/app-event-type.enum.ts
export enum AppEventType {
  // --- Propuestas ---
  PROPOSAL_CREATED = 'PROPOSAL_CREATED',
  EVALUATION_ASSIGNED = 'EVALUATION_ASSIGNED',
  PROPOSAL_CORRECTION_UPLOADED = 'PROPOSAL_CORRECTION_UPLOADED',
  PROPOSAL_DEADLINE_WARNING = 'PROPOSAL_DEADLINE_WARNING',
  PROPOSAL_DEADLINE_EXPIRED = 'PROPOSAL_DEADLINE_EXPIRED',
  // --- Anteproyecto ---
  PRELIMINARY_DRAFT_CREATED = 'PRELIMINARY_DRAFT_CREATED',
  REVIEWERS_ASSIGNED = 'REVIEWERS_ASSIGNED',
  PRELIMINARY_DRAFT_EVALUATION_REGISTERED = 'PRELIMINARY_DRAFT_EVALUATION_REGISTERED',
  PRELIMINARY_DRAFT_CORRECTION_UPLOADED = 'PRELIMINARY_DRAFT_CORRECTION_UPLOADED',
  PRELIMINARY_DRAFT_COUNCIL_PRESENTATION_UPLOADED = 'PRELIMINARY_DRAFT_COUNCIL_PRESENTATION_UPLOADED',
  COUNCIL_RESOLUTION_UPLOADED = 'COUNCIL_RESOLUTION_UPLOADED',
  PRELIMINARY_DRAFT_DEADLINE_WARNING = 'PRELIMINARY_DRAFT_DEADLINE_WARNING',
  PRELIMINARY_DRAFT_DEADLINE_EXPIRED = 'PRELIMINARY_DRAFT_DEADLINE_EXPIRED',
  // --- Trabajos de Grado ---
  THESIS_ADVANCE_UPLOADED = 'THESIS_ADVANCE_UPLOADED',
  THESIS_ADVANCE_EVALUATED = 'THESIS_ADVANCE_EVALUATED',
  THESIS_FINAL_DELIVERY_UPLOADED = 'THESIS_FINAL_DELIVERY_UPLOADED',
  THESIS_SUSTENTATION_PROGRAMMED = 'THESIS_SUSTENTATION_PROGRAMMED',
  THESIS_VERDICT_REGISTERED = 'THESIS_VERDICT_REGISTERED',
  SPECIAL_REQUEST_CREATED = 'SPECIAL_REQUEST_CREATED',
  SPECIAL_REQUEST_RESOLVED = 'SPECIAL_REQUEST_RESOLVED',
  THESIS_PAZ_Y_SALVO_REGISTERED = 'THESIS_PAZ_Y_SALVO_REGISTERED',
  THESIS_CORRECTED_DOCUMENTS_UPLOADED = 'THESIS_CORRECTED_DOCUMENTS_UPLOADED',
  THESIS_CORRESPONDENCE_REGISTERED = 'THESIS_CORRESPONDENCE_REGISTERED',
  THESIS_CORRECTED_DOCUMENTS_EVALUATED = 'THESIS_CORRECTED_DOCUMENTS_EVALUATED',
  THESIS_REACTIVATED = 'THESIS_REACTIVATED',
  THESIS_DEADLINE_WARNING = 'THESIS_DEADLINE_WARNING',
  THESIS_DEADLINE_EXPIRED = 'THESIS_DEADLINE_EXPIRED'
}